Fix maas apiclient to work with MAASv2 api
Also embedded the api client in the code rather than syncing with
maasclient. This is a simple class to wrap get/post/put/delete requests
against the maas server. The querying is left up to the developer, for
example,

from conjureup.maas import MaasClient
m = MaasClient(server_address='127.0.0.1', consumer_key='...')
res = m.get('/nodes/')
res.ok
